+++ src/researchforge/__init__.py
@@ -1,0 +1,3 @@
+"""ResearchForge application package."""
+
+__version__ = "0.1.0"